+++ d31337m3/microservices/cityhall/app/routers/auth.py
@@ -7,7 +7,7 @@ import urllib.error
 import json
 
 from fastapi import APIRouter, Depends, HTTPException, Request, status
-from pydantic import BaseModel, EmailStr
+from pydantic import BaseModel, EmailStr, Field
 from sqlalchemy import select, update
 from sqlalchemy.ext.asyncio import AsyncSession